requirements.txt
# Core FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.9
pydantic==2.5.0

# HTTP client and retries
httpx==0.25.2
tenacity==8.2.3

# Trading and market data
alpaca-trade-api==3.1.1
pandas==2.1.3
numpy==1.25.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.14.0
xgboost==2.0.1

# Configuration and utilities
python-dotenv==1.0.0

# Data processing
pytz==2023.3
requests==2.31.0

# System monitoring (optional)
psutil==5.9.6

# Development tools (optional)
jupyter==1.0.0

# Testing and Development Dependencies
pytest==8.4.1
pytest-asyncio==1.1.0
respx==0.22.0
anyio==4.8.0
ruff==0.12.8
mypy==1.17.1
